S4 g1 refactor siena

+++ server/express.js
@@ -88,6 +88,7 @@ con.connect(function(err) {
     let sql = 'INSERT IGNORE INTO items (vendor_name, descriptor, req_department, item_desc, unit_price, dep_desc, item_total, product_name, po_no, entry_id, issue_date, vendor_code, po_quantity) VALUES ?'
     // let sql = 'INSERT INTO items (vendor_name, descriptor, req_department, item_desc, unit_price, dep_desc, item_total, product_name, po_no, entry_id, issue_date, vendor_code) VALUES ?'
     const userPass = 'INSERT IGNORE INTO users VALUES (\'scotty@ucr.edu\', \'thebear\', \'riverside\')'
+    
     con.query(sql, [values], function(err,result) {
       if(err) {
         throw err;
@@ -149,8 +150,10 @@ app.post('/signIn', (req, res) => {
   con.query(sqlquery, function (err, result) {
     if (err) throw err;
     //results = (result[0].cnt)
-    console.log(result)
-    res.status(200).json(results)
+    if (result.length == 0) console.log('it is empty')
+
+
+    res.status(200).json(result)
   });
 })
 
@@ -202,33 +205,46 @@ app.get('/serverUpload', (req,res) => {
 let results = [];
 app.post('/serverUpload', (req, res, next) => {
   const file = req.files.file;
+  let toNLPfile = path.join(__dirname + './ToNLP/newData.json')
+  fp = path.join(__dirname + '/../uploads/', file.name)
+  
+  if (fs.existsSync(fp)) {
+    res.send({
+      success: false,
+      message: 'File Already Exists'
+    });
+  }
+  else {
+    file.mv("./uploads/" + file.name, function(err, result) {
+      if (err) throw err;
+      let [fileName, fileExtension] = (file.name).split('.')
 
-  file.mv("./uploads/" + file.name, function(err, result) {
-    if (err) throw err;
-    fp = path.join(__dirname + '/../uploads/', file.name)
-    let [fileName, fileExtension] = (file.name).split('.')
-    delayedRes = async () => { 
-      fs.createReadStream(fp)
-      .on('error', () => {
-          console.log("errorr")
-      })
-      .pipe(csv())
-      .on('data', (row) => {
-          results.push(row)
-      })
-      .on('end', () => {
-        fileName = './server/ToNLP/' + fileName + '.json'
-        fs.writeFile(fileName,  JSON.stringify(results), function (err) {
-          if (err) throw err;
-          console.log('Saved!');
-        });
-      })
-    }
-    delayedRes()
-
+      delayedRes = async() => { 
+        fs.createReadStream(fp)
+        .on('error', () => {
+            console.log("errorr")
+        })
+        .pipe(csv())
+        .on('data', (row) => {
+            results.push(row)
+        })
+        .on('end', () => {
+          fileName = './server/ToNLP/newData.json'
+          if (fs.existsSync(toNLPfile)) {
+            console.log("UNLINKINGGGG")
+            fs.unlinkSync(toNLPfile)
+          }
+          fs.writeFile(fileName, JSON.stringify(results), function (err) {
+            if (err) throw err;
+            console.log('Saved!');
+          });
+        })
+      }
+      delayedRes()
+    })
     res.send({
       success: true,
-      message: 'file Uploaded'
+      message: 'Successfuly Uploaded'
     });
-  });
+  }
 })
